build: opt code of build each & move dialog style import to dialog.js in es
