--- packages/web-vue/script/common.mts
@@ -1,21 +1,0 @@
-export const outputGlobal = {
-	vue: 'Vue',
-	'@floating-ui/dom': 'FloatingUI',
-	overlayscrollbars: 'OverlayScrollbarsGlobal',
-	'overlayscrollbars-vue': 'OverlayScrollbarsVue',
-	'parsnip-kit': 'parsnip-kit',
-	'vue-router': 'VueRouter',
-	mitt: 'mitt'
-}
-
-export const target = ['es2020', 'edge88', 'firefox78', 'chrome87', 'safari14']
-
-export const external = [
-	'vue',
-	'@floating-ui/dom',
-	'overlayscrollbars',
-	'overlayscrollbars-vue',
-	'parsnip-kit',
-	'vue-router',
-	'mitt'
-]
